2/23/14-2
Fixed up the board so that it is a 2d array now.  x and y
positions(0-9) are set when each button is created.  each button now
set up to accept left and right mouse clicks.  next step for me-set
different button text when right click is pressed.  Now most likely
done for day.